the-one-brain-main/backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==1.10.15
pydantic-settings==2.0.3
python-multipart==0.0.6
sqlmodel==0.0.14
pillow==10.4.0
anthropic>=0.25.0
python-dotenv==1.0.0
